requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0

# CLI dependencies
click>=8.0.0
rich>=13.0.0
pyyaml>=6.0.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0
black>=22.0.0
mypy>=1.0.0
isort>=5.0.0